nodes/dm-and/pyproject.toml
[project]
name = "dm-and"
version = "0.1.0"
description = "Logic node that emits true only when all configured inputs are true"
requires-python = ">=3.10"
dependencies = [
  "dora-rs>=0.4.1,<0.5.0",
  "pyarrow>=21.0.0",
]

[project.scripts]
dm-and = "dm_and.main:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
